requirements-secure.txt
# Requirements con versiones seguras - Sin vulnerabilidades conocidas
# Actualizado: 26 de Noviembre de 2025
# 
# Cambios respecto a requirements.txt:
# - gunicorn: 21.2.0 → 23.0.0 (Corrige CVE-2024-1135, CVE-2024-6827)
# - Jinja2: 3.1.4 → 3.1.6 (Corrige CVE-2024-56201, CVE-2025-27516, CVE-2024-56326)
# - Werkzeug: 3.0.0 → 3.1.3 (Corrige CVE-2023-46136, CVE-2024-49766, CVE-2024-49767, CVE-2024-34069, y otros)

Flask==3.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
bcrypt==4.1.1
flask-wtf==1.2.0
wtforms==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.0
prometheus-client==0.19.0
